requirements.txt
Kivy==2.3.0
pyjnius==1.7.0
sentence_transformers==3.0.1
torch==2.3.1
transformers==4.41.2
uiautomator2==3.4.1
xmltodict==0.13.0
